+++ sample/RazorApp/Startup.cs
@@ -7,8 +7,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-using Wangkanai.Responsive;
-
 namespace RazorApp
 {
     public class Startup
