+++ src/DependencyInjection/ApplicationBuilderExtensions.cs
@@ -5,7 +5,7 @@ using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Logging;
 using Wangkanai.Detection.DependencyInjection.Options;
-using Wangkanai.Detection.Internal;
+using Wangkanai.Detection.Services;
 using Wangkanai.Detection.Responsive;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -41,7 +41,7 @@ namespace Microsoft.Extensions.DependencyInjection
                 throw new ArgumentNullException(nameof(loggerFactory));
 
             var logger = loggerFactory.CreateLogger("Detection.Startup");
-            logger.LogInformation("Starting Detection version {version}", typeof(DetectionApplicationBuilderExtensions).Assembly.GetName().Version.ToString());
+            logger.LogInformation("Starting Detection version {version}", typeof(DetectionApplicationBuilderExtensions)?.Assembly?.GetName()?.Version?.ToString());
 
             var scopeFactory = app.ApplicationServices.GetService<IServiceScopeFactory>();
