+++ src/DependencyInjection/BuilderExtensions/CoreCollectionExtensions.cs
@@ -4,7 +4,6 @@
 using System;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Wangkanai.Detection.Internal;
 using Wangkanai.Detection.Services;
 
 namespace Microsoft.Extensions.DependencyInjection
