--- src/Internal/DetectionMarkerService.cs
@@ -1,13 +1,0 @@
-// Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
-// The Apache v2. See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Wangkanai.Detection.Internal
-{
-    internal class DetectionMarkerService
-    {
-    }
-}
