+++ src/Services/DetectionMarkerService.cs
@@ -1,9 +1,9 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
-namespace Wangkanai.Responsive.Internal
+namespace Wangkanai.Detection.Services
 {
-    internal class ResponsiveMarkerService
+    internal class DetectionMarkerService
     {
     }
 }
