+++ test/Client.cs
@@ -8,6 +8,6 @@ namespace Wangkanai.Detection
         //public Platform Platform { get; set; }
         public Engine Engine { get; set; }
 
-        public string? Agent { get; set; }
+        public string Agent { get; set; }
     }
 }
