+++ test/Device/DeviceTabletTest.cs
@@ -1,7 +1,6 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
-using Wangkanai.Detection.Test;
 using Xunit;
 
 namespace Wangkanai.Detection.DeviceTest
